+++ pygmo/island.cpp
@@ -9,6 +9,8 @@
 #include <memory>
 #include <sstream>
 #include <string>
+#include <typeindex>
+#include <typeinfo>
 #include <vector>
 
 #include <boost/numeric/conversion/cast.hpp>
@@ -16,6 +18,7 @@
 #include <pybind11/pybind11.h>
 
 #include <pagmo/algorithm.hpp>
+#include <pagmo/config.hpp>
 #include <pagmo/detail/make_unique.hpp>
 #include <pagmo/island.hpp>
 #include <pagmo/population.hpp>
@@ -127,6 +130,25 @@ std::string isl_inner<py::object>::get_extra_info() const
 {
     return getter_wrapper<std::string>(m_value, "get_extra_info", std::string{});
 }
+
+#if PAGMO_VERSION_MAJOR > 2 || (PAGMO_VERSION_MAJOR == 2 && PAGMO_VERSION_MINOR >= 15)
+
+std::type_index isl_inner<py::object>::get_type_index() const
+{
+    return std::type_index(typeid(py::object));
+}
+
+const void *isl_inner<py::object>::get_ptr() const
+{
+    return &m_value;
+}
+
+void *isl_inner<py::object>::get_ptr()
+{
+    return &m_value;
+}
+
+#endif
 
 template <typename Archive>
 void isl_inner<py::object>::save(Archive &ar, unsigned) const