+++ pygmo/problem.cpp
@@ -10,6 +10,8 @@
 #include <memory>
 #include <sstream>
 #include <string>
+#include <typeindex>
+#include <typeinfo>
 #include <utility>
 #include <vector>
 
@@ -19,6 +21,7 @@
 #include <pybind11/numpy.h>
 #include <pybind11/pybind11.h>
 
+#include <pagmo/config.hpp>
 #include <pagmo/detail/make_unique.hpp>
 #include <pagmo/problem.hpp>
 #include <pagmo/s11n.hpp>
@@ -345,6 +348,25 @@ thread_safety prob_inner<py::object>::get_thread_safety() const
 {
     return thread_safety::none;
 }
+
+#if PAGMO_VERSION_MAJOR > 2 || (PAGMO_VERSION_MAJOR == 2 && PAGMO_VERSION_MINOR >= 15)
+
+std::type_index prob_inner<py::object>::get_type_index() const
+{
+    return std::type_index(typeid(py::object));
+}
+
+const void *prob_inner<py::object>::get_ptr() const
+{
+    return &m_value;
+}
+
+void *prob_inner<py::object>::get_ptr()
+{
+    return &m_value;
+}
+
+#endif
 
 template <typename Archive>
 void prob_inner<py::object>::save(Archive &ar, unsigned) const