+++ pygmo/topology.cpp
@@ -11,6 +11,8 @@
 #include <memory>
 #include <sstream>
 #include <string>
+#include <typeindex>
+#include <typeinfo>
 #include <utility>
 #include <vector>
 
@@ -19,6 +21,7 @@
 #include <pybind11/numpy.h>
 #include <pybind11/pybind11.h>
 
+#include <pagmo/config.hpp>
 #include <pagmo/s11n.hpp>
 #include <pagmo/topology.hpp>
 #include <pagmo/types.hpp>
@@ -137,6 +140,38 @@ std::string topo_inner<py::object>::get_extra_info() const
 {
     return getter_wrapper<std::string>(m_value, "get_extra_info", std::string{});
 }
+
+#if PAGMO_VERSION_MAJOR > 2 || (PAGMO_VERSION_MAJOR == 2 && PAGMO_VERSION_MINOR >= 15)
+
+bgl_graph_t topo_inner<py::object>::to_bgl() const
+{
+    auto m = pygmo::callable_attribute(m_value, "to_networkx");
+    if (m.is_none()) {
+        pygmo::py_throw(PyExc_NotImplementedError,
+                        ("the to_networkx() conversion method has been invoked in the user-defined Python topology '"
+                         + pygmo::str(m_value) + "' of type '" + pygmo::str(pygmo::type(m_value))
+                         + "', but the method is either not present or not callable")
+                            .c_str());
+    }
+    return pygmo::networkx_to_bgl_graph_t(m());
+}
+
+std::type_index topo_inner<py::object>::get_type_index() const
+{
+    return std::type_index(typeid(py::object));
+}
+
+const void *topo_inner<py::object>::get_ptr() const
+{
+    return &m_value;
+}
+
+void *topo_inner<py::object>::get_ptr()
+{
+    return &m_value;
+}
+
+#endif
 
 template <typename Archive>
 void topo_inner<py::object>::save(Archive &ar, unsigned) const