+++ pygmo/topology.hpp
@@ -13,11 +13,13 @@
 #include <memory>
 #include <string>
 #include <type_traits>
+#include <typeindex>
 #include <utility>
 #include <vector>
 
 #include <pybind11/pybind11.h>
 
+#include <pagmo/config.hpp>
 #include <pagmo/s11n.hpp>
 #include <pagmo/topology.hpp>
 #include <pagmo/types.hpp>
@@ -53,6 +55,14 @@ struct topo_inner<py::object> final : topo_inner_base, pygmo::common_base {
     // Optional methods.
     virtual std::string get_name() const override final;
     virtual std::string get_extra_info() const override final;
+
+#if PAGMO_VERSION_MAJOR > 2 || (PAGMO_VERSION_MAJOR == 2 && PAGMO_VERSION_MINOR >= 15)
+    virtual bgl_graph_t to_bgl() const override final;
+    virtual std::type_index get_type_index() const override final;
+    virtual const void *get_ptr() const override final;
+    virtual void *get_ptr() override final;
+#endif
+
     template <typename Archive>
     void save(Archive &, unsigned) const;
     template <typename Archive>