+++ tools/travis_osx.sh
@@ -13,9 +13,9 @@ bash miniconda.sh -b -p $HOME/miniconda
 conda config --add channels conda-forge --force
 
 if [[ "${PYGMO_BUILD_TYPE}" == *pagmo_head ]]; then
-    conda_pkgs="cmake eigen nlopt ipopt boost-cpp tbb tbb-devel python=${PYTHON_VERSION} numpy cloudpickle dill numba pybind11 clang clangdev ipyparallel scipy"
+    conda_pkgs="cmake eigen nlopt ipopt boost-cpp tbb tbb-devel python=${PYTHON_VERSION} numpy cloudpickle networkx dill numba pybind11 clang clangdev ipyparallel"
 else
-    conda_pkgs="cmake boost-cpp python=${PYTHON_VERSION} numpy cloudpickle dill numba pybind11 clang clangdev ipyparallel pagmo-devel scipy"
+    conda_pkgs="cmake boost-cpp python=${PYTHON_VERSION} numpy cloudpickle networkx dill numba pybind11 clang clangdev ipyparallel pagmo-devel"
 fi
 conda create -q -p $deps_dir -y
 source activate $deps_dir